+++ src/Constants.py
@@ -1,0 +1,2 @@
+DATABASE_NAME = 'severus'
+IMAGE_DIR_PATH = '/home/manisha/Work/Severus/drive-download-20220409T162323Z-001'
